+++ packages/seroval/src/core/cross/parser.ts
