+++ packages/seroval/src/core/plugin.ts
@@ -2,7 +2,7 @@ import type BaseAsyncParserContext from './context/parser/async';
 import type BaseStreamParserContext from './context/parser/stream';
 import type BaseSyncParserContext from './context/parser/sync';
 import type BaseSerializerContext from './context/serializer';
-import type VanillaDeserializerContext from './tree/deserialize';
+import type BaseDeserializerContext from './context/deserializer';
 
 export type SerovalMode = 'vanilla' | 'cross';
 
@@ -37,7 +37,7 @@ export interface Plugin<Value, Node> {
   ): string;
   deserialize(
     node: Node,
-    ctx: VanillaDeserializerContext,
+    ctx: BaseDeserializerContext,
     data: PluginData,
   ): Value;
   // isIterable?: (value: Value) => boolean;
