+++ packages/seroval/src/core/tree/serializer.ts
